Float change

+++ args.py
@@ -27,7 +27,7 @@ def init_args():
     parser.add_argument('--super_clusters', type=int, default=20)
     parser.add_argument('--freeze_losers', type=lambda x: (str(x).lower() == 'true'), default=False)
     #Similarity Experiment
-    parser.add_argument('--k_fraction', type=int, default=0.25)
+    parser.add_argument('--k_fraction', type=float, default=0.25)
 
 
     parser.add_argument(
